fix(package.json): remove unnecessary script part
I do not want to run scripts, this is a really simple project.
If the project will be complex then probably I will write
scripts.

If you need to perform operations on your package before it is
used, in a way that is not dependent on the operating system or
architecture of the target system, use a prepublish script.

This includes tasks such as:

- Compiling CoffeeScript source code into JavaScript.
- Creating minified versions of JavaScript source code.
- Fetching remote resources that your package will use.

Signed-off-by: Ádám Gólya <adam.golya@lab.coop>