feat(server.js): initial add of server.js
This file will be the only entry point for the server side, we do not
want to start other service for the REST API with oAuth2.

We need to figure out what will be the place of the middlewares, in
the future the server.js file can not contain middleware.

Signed-off-by: Ádám Gólya <adam.golya@lab.coop>